Rename drivers/drivers to drivers/index
